multitenant: add SQL server startup guardrails
This code change prevents a SQL server from starting if its
binary version is less than the tenant's active version.

It moves the decoder of the settingswatcher under the mutex
to synchronize its usage and prevent data races because we now
use it in multiple places.

The check runs after we run permanent upgrades because
that's when the version setting is written to the settings table
of the system tenant (and we want to apply this check to both system
and secondary tenants). To guard against future changes to where we
set the version, if GetTenantClusterVersion doesn't find a value for
the version setting it will fatal because this means the
version is not set yet.

Release note: None
Epic: CRDB-10829

+++ pkg/server/server_sql.go
@@ -1483,6 +1483,18 @@ func (s *SQLServer) preStart(
 
 	log.Infof(ctx, "done ensuring all necessary startup migrations have run")
 
+	// Prevent the server from starting if its binary version is too low
+	// for the current tenant cluster version.
+	// This check needs to run after the "version" setting is set in the
+	// "system.settings" table of this tenant. This includes both system
+	// and secondary tenants.
+	tenantActiveVersion := s.settingsWatcher.GetTenantClusterVersion(ctx, s.execCfg.DB).Version
+	if s.execCfg.Settings.Version.BinaryVersion().Less(tenantActiveVersion) {
+		return errors.Newf("preventing SQL server from starting because its binary version "+
+			"is too low for the tenant active version: server binary version = %v, tenant active version = %v",
+			s.execCfg.Settings.Version.BinaryVersion(), tenantActiveVersion)
+	}
+
 	// Delete all orphaned table leases created by a prior instance of this
 	// node. This also uses SQL.
 	s.leaseMgr.DeleteOrphanedLeases(ctx, orphanedLeasesTimeThresholdNanos)
